Updated the following packages: xserver          commit 08113b8923a83aa7018463e71b6e2db3661b8604 libxcb           commit 622152ee42a310876f10602601206954b8d0613e libxcb/xcb-proto commit 77d7fc04da729ddc5ed4aacf30253726fac24dca xkeyboard-config commit 022c3ca1db12e505cbd5ce8bf19c432d6a70c7e5 libX11           commit a465588218c1643eedc35b3c24409cb775454eee libXdmcp         commit 1192d3bc407348ff316bd3bffc791b3ac73f591b libXext          commit 424b67ad03fafdb0d6caf2e9bf5a103e9717e71f xkbcomp          commit 0b4873ce55e3afe2e1716ef4a21cd94a7ae96e64 pixman           commit b252d40714bc23d6201a93d599a23193a2875468 mkfontscale      commit bcfa27aeef3a9eb720e6b48e3fb8a4c2813fef52 libXft           commit e736f48ba2fd4263ca6f692e27916a914447af4e libXmu           commit 291feb3d6d70758185de71a21cf87d23a4b965cd fontconfig       commit 3b4641ae7a7bf961221cd7b55cc0e25d9d1a6365 mesa             commit cc9141f0449b84204172f5729ea3dbc1724239c0 putty            commit 6b10eaa245f05c938085d61e9de4cc27d5bb1611 pthreads         commit 8e467a62a14fd9de15af33beec0a913d23f4d2f9 xorgproto        commit 68de489ec6c2fb6f8cfc47b0bba7edd0f9942f17

+++ mesalib/src/freedreno/fdl/freedreno_layout.h
@@ -113,8 +113,8 @@ struct fdl_layout {
    struct fdl_slice ubwc_slices[FDL_MAX_MIP_LEVELS];
    uint32_t pitch0;
    uint32_t ubwc_width0;
-   uint32_t layer_size;
-   uint32_t ubwc_layer_size; /* in bytes */
+   uint64_t layer_size;
+   uint64_t ubwc_layer_size; /* in bytes */
    bool ubwc : 1;
    bool layer_first : 1; /* see above description */
    bool tile_all : 1;
@@ -143,7 +143,7 @@ struct fdl_layout {
    uint32_t nr_samples;
    enum pipe_format format;
 
-   uint32_t size;       /* Size of the whole image, in bytes. */
+   uint64_t size;       /* Size of the whole image, in bytes. */
    uint32_t base_align; /* Alignment of the base address, in bytes. */
    uint8_t pitchalign;  /* log2(pitchalign) */
 };
